+++ Opcodes/hrtferx.h
@@ -31,7 +31,7 @@ typedef struct {
   MEMFIL        *mfp;      /* file pointer */
   int16         *fpbegin;
   int           oldel_index, oldaz_index;
-  long          incount, outfront, outend, outcount;
+  int32         incount, outfront, outend, outcount;
   AUXCH         auxch;      /* will point to allocated memory */
   HRTF_DATUM    hrtf_data, oldhrtf_data;  /* matrix to store HRTF data */
   MYFLT         outl[BUF_LEN], outr[BUF_LEN];
